fixing errors

+++ src/main/java/org/example/technologie_sieciowe_1/service/BookDetailsService.java
@@ -29,8 +29,8 @@ public class BookDetailsService {
         return bookDetailsRepository.save(book);
     }
 
-    public void delete(Long id) {
-        bookDetailsRepository.deleteById(Math.toIntExact(id));
+    public void delete(Integer id) {
+        bookDetailsRepository.deleteById(id);
     }
 
 }
